put files in the right places
